set cuda int8 as default inference mode

+++ src/aura/config.py
@@ -5,7 +5,7 @@ VAD_LEVEL = 3
 
 MODEL_ID = "SoybeanMilk/faster-whisper-Breeze-ASR-25"
 DEVICE = "cuda"
-COMPUTE_TYPE = "float16"
+COMPUTE_TYPE = "int8"
 GITHUB_REPOSITORY = "JasonLn0711/project_aura"
 
 DEFAULT_PROMPT = "這是一份專業的繁體中文會議紀錄，請務必根據語氣加上正確的全形標點符號。"
